Split up our test actions into unit, integ, and doctest. (#4193)
* Run unit test suites in parallel


* Split out our test actions


* Make unit test step run in parallel


* Fix removed bwc tests


* Add another missing parallel flag


---------
