+++ gradle.properties
@@ -4,3 +4,4 @@
 
 version=1.13.0
 org.gradle.jvmargs=-Duser.language=en -Duser.country=US
+org.gradle.parallel=true
